+++ docs/CHANGELOG.md
@@ -1,3 +1,4 @@
+- #1446 Integrate stf_info_manager in runner.rs
 - #1450 Renames two related traits: `RuntimeAuthorization` becomes `TransactionAuthorizer`; `RuntimeAuthentiation` beomces `TransactionAuthenticator`. The `runtime_authorization` method on the `HasCapabilities` trait is also renamed to `transaction_authorizer`. 
 - #1443 Move StateTransitionInfo channel creation to `FullNodeBlueprint`.
 - #1442 Makes stf-runner handle trailing slashes without throwing HTTP 404.